+++ webextension/js/init-templates.js
@@ -12,7 +12,6 @@ export function nunjuckRender(templateName, context) {
 
 
 const templatesSource = window.templatesSource = new Map();
-templatesSource.set('tabMover', '/templates/panel/tabMover');
 templatesSource.set('newTab', '/templates/newTab');